server/requirements.txt
numpy==1.19.2
pandas==1.1.3
scikit-learn==0.23.2
matplotlib==3.3.2
xai-sdk
tweetpy
python-dotenv